Fix forward declarations in publicly facing osrm header
